Make brand buttons reusable
